requirements.txt
streamlit~=1.29.0
pandas~=2.0.0
neo4j~=5.19.0
setuptools~=67.7.1
networkx~=3.3
Pillow~=9.0.1
pydicom~=2.3.1
numpy~=1.24.2
matplotlib~=3.7.1
tqdm~=4.66.4
neomodel~=5.3.0
tifffile~=2024.9.20
PyYAML~=5.4.1